Metadesc/header/other small changes (#815)
I see that our metatitle for this is "Solana (SOL) Faucet | thirdweb" - it used to show up like this on Google, but it seems to be showing as "Solana Faucet" now for some reason. (can't attach screenshot but tried on different windows/browsers)

Because of this, we're missing out on a lot of traffic that goes to "sol faucet" rather than "solana faucet" - we rank #4 for "solana faucet" but not at all for "sol faucet" & could easily rank in the top 5 for sol faucet as well!

Generally seeing discrepancies across Dashboard pages between what shows up on Google, what it shows up as when shared (og:title), and what shows up when on-page. @jnsdls hope you feel better throughout the week - lmk if you'll be able to make it to our SEO biweekly or if ya wanna push!